Split out cluster/install functionality into more reasonable locations
