+++ pkg/cluster/cluster.go
@@ -1,0 +1,127 @@
+package cluster
+
+// Copyright (c) Microsoft Corporation.
+// Licensed under the Apache License 2.0.
+
+import (
+	"context"
+
+	"github.com/Azure/go-autorest/autorest/azure"
+	configclient "github.com/openshift/client-go/config/clientset/versioned"
+	operatorclient "github.com/openshift/client-go/operator/clientset/versioned"
+	samplesclient "github.com/openshift/client-go/samples/clientset/versioned"
+	securityclient "github.com/openshift/client-go/security/clientset/versioned"
+	"github.com/sirupsen/logrus"
+	extensionsclient "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset"
+	"k8s.io/client-go/kubernetes"
+
+	"github.com/Azure/ARO-RP/pkg/api"
+	"github.com/Azure/ARO-RP/pkg/database"
+	"github.com/Azure/ARO-RP/pkg/env"
+	aroclient "github.com/Azure/ARO-RP/pkg/operator/clientset/versioned/typed/aro.openshift.io/v1alpha1"
+	"github.com/Azure/ARO-RP/pkg/util/azureclient/mgmt/compute"
+	"github.com/Azure/ARO-RP/pkg/util/azureclient/mgmt/features"
+	"github.com/Azure/ARO-RP/pkg/util/azureclient/mgmt/network"
+	"github.com/Azure/ARO-RP/pkg/util/azureclient/mgmt/storage"
+	"github.com/Azure/ARO-RP/pkg/util/billing"
+	"github.com/Azure/ARO-RP/pkg/util/dns"
+	"github.com/Azure/ARO-RP/pkg/util/encryption"
+	"github.com/Azure/ARO-RP/pkg/util/keyvault"
+	"github.com/Azure/ARO-RP/pkg/util/privateendpoint"
+	"github.com/Azure/ARO-RP/pkg/util/refreshable"
+	"github.com/Azure/ARO-RP/pkg/util/subnet"
+)
+
+// manager contains information needed to install and maintain an ARO cluster
+type manager struct {
+	log             *logrus.Entry
+	env             env.Interface
+	db              database.OpenShiftClusters
+	billing         billing.Manager
+	doc             *api.OpenShiftClusterDocument
+	subscriptionDoc *api.SubscriptionDocument
+	cipher          encryption.Cipher
+	fpAuthorizer    refreshable.Authorizer
+
+	disks             compute.DisksClient
+	virtualmachines   compute.VirtualMachinesClient
+	vnet              network.VirtualNetworksClient
+	interfaces        network.InterfacesClient
+	publicipaddresses network.PublicIPAddressesClient
+	loadbalancers     network.LoadBalancersClient
+	securitygroups    network.SecurityGroupsClient
+	deployments       features.DeploymentsClient
+	groups            features.ResourceGroupsClient
+	accounts          storage.AccountsClient
+
+	dns             dns.Manager
+	keyvault        keyvault.Manager
+	privateendpoint privateendpoint.Manager
+	subnet          subnet.Manager
+
+	kubernetescli kubernetes.Interface
+	extcli        extensionsclient.Interface
+	operatorcli   operatorclient.Interface
+	configcli     configclient.Interface
+	samplescli    samplesclient.Interface
+	securitycli   securityclient.Interface
+	arocli        aroclient.AroV1alpha1Interface
+}
+
+const deploymentName = "azuredeploy"
+
+// New returns a cluster manager
+func New(ctx context.Context, log *logrus.Entry, _env env.Interface, db database.OpenShiftClusters,
+	billing billing.Manager, doc *api.OpenShiftClusterDocument, subscriptionDoc *api.SubscriptionDocument) (*manager, error) {
+	r, err := azure.ParseResourceID(doc.OpenShiftCluster.ID)
+	if err != nil {
+		return nil, err
+	}
+
+	localFPAuthorizer, err := _env.FPAuthorizer(_env.TenantID(), azure.PublicCloud.ResourceManagerEndpoint)
+	if err != nil {
+		return nil, err
+	}
+
+	localFPKVAuthorizer, err := _env.FPAuthorizer(_env.TenantID(), azure.PublicCloud.ResourceIdentifiers.KeyVault)
+	if err != nil {
+		return nil, err
+	}
+
+	fpAuthorizer, err := _env.FPAuthorizer(doc.OpenShiftCluster.Properties.ServicePrincipalProfile.TenantID, azure.PublicCloud.ResourceManagerEndpoint)
+	if err != nil {
+		return nil, err
+	}
+
+	cipher, err := encryption.NewXChaCha20Poly1305(ctx, _env, env.EncryptionSecretName)
+	if err != nil {
+		return nil, err
+	}
+
+	return &manager{
+		log:             log,
+		env:             _env,
+		db:              db,
+		billing:         billing,
+		cipher:          cipher,
+		doc:             doc,
+		subscriptionDoc: subscriptionDoc,
+		fpAuthorizer:    fpAuthorizer,
+
+		disks:             compute.NewDisksClient(r.SubscriptionID, fpAuthorizer),
+		virtualmachines:   compute.NewVirtualMachinesClient(r.SubscriptionID, fpAuthorizer),
+		vnet:              network.NewVirtualNetworksClient(r.SubscriptionID, fpAuthorizer),
+		interfaces:        network.NewInterfacesClient(r.SubscriptionID, fpAuthorizer),
+		publicipaddresses: network.NewPublicIPAddressesClient(r.SubscriptionID, fpAuthorizer),
+		loadbalancers:     network.NewLoadBalancersClient(r.SubscriptionID, fpAuthorizer),
+		securitygroups:    network.NewSecurityGroupsClient(r.SubscriptionID, fpAuthorizer),
+		deployments:       features.NewDeploymentsClient(r.SubscriptionID, fpAuthorizer),
+		groups:            features.NewResourceGroupsClient(r.SubscriptionID, fpAuthorizer),
+		accounts:          storage.NewAccountsClient(r.SubscriptionID, fpAuthorizer),
+
+		dns:             dns.NewManager(_env, localFPAuthorizer),
+		keyvault:        keyvault.NewManager(localFPKVAuthorizer),
+		privateendpoint: privateendpoint.NewManager(_env, localFPAuthorizer),
+		subnet:          subnet.NewManager(r.SubscriptionID, fpAuthorizer),
+	}, nil
+}
